requirements/dev.txt
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
cryptography==41.0.3
dataclasses==0.6
defusedxml==0.7.1
deprecation==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
gotrue==1.0.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
postgrest==0.10.8
postgrest-py==0.4.0
psycopg2-binary==2.9.7
py==1.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3
realtime==1.0.0
realtime-py==0.1.3
requests==2.31.0
requests-oauthlib==1.3.1
resend==0.5.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supabase-py==0.0.2
supafunc==0.2.3
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.16
websockets==10.4
